+++ backend/src/stems/BsRoformerRhythm.cpp
@@ -3,7 +3,6 @@
 #include <algorithm>
 #include <array>
 #include <cmath>
-#include <thread>
 #include <vector>
 
 #include <onnxruntime_cxx_api.h>
@@ -13,6 +12,7 @@
 #endif
 
 #include "BsRoformerSpectral.h"
+#include "InferenceThreads.h"
 #include "Log.h"
 #include "StemSeparator.h"
 
@@ -43,8 +43,7 @@ struct BsRoformerRhythm::Impl
         session.reset();
         sessionPath = {};
         sessionOptions = Ort::SessionOptions{};
-        const unsigned int cores = std::thread::hardware_concurrency();
-        sessionOptions.SetIntraOpNumThreads(static_cast<int>(std::max(1u, cores)));
+        sessionOptions.SetIntraOpNumThreads(stems::inferenceIntraOpThreads());
         sessionOptions.SetGraphOptimizationLevel(GraphOptimizationLevel::ORT_ENABLE_ALL);
         if (useGpu)
         {